+++ AIPScannerConfig/functions/New-AIPScannerInstall.ps1
@@ -72,7 +72,9 @@
                     Write-PSFMessage -Level Verbose -String 'New-AIPScannerInstall.Message7' -StringValues $aipScannerModule.Name
                     throw "Module import failed!"
                 }
+            }
 
+            try {
                 Write-PSFMessage -Level Verbose -String 'New-AIPScannerInstall.Message8'
                 Install-AIPScanner -SqlServerInstance "$env:ComputerName\SQLExpress" -ErrorAction Stop
 
@@ -80,42 +82,45 @@
                     $Cluster = Read-Host "Please enter your Azure AIP Cluster Name"
                 }
             }
+            catch {
+                Stop-PSFFunction -String 'New-AIPScannerInstall.Message9' -EnableException $EnableException -Cmdlet $PSCmdlet -ErrorRecord $_
+            }
 
             try {
-                Write-PSFMessage -Level Verbose -String 'New-AIPScannerInstall.Message9'
+                Write-PSFMessage -Level Verbose -String 'New-AIPScannerInstall.Message10'
                 if (Set-AIPScanner -SqlServerInstance "$env:COMPUTERNAME\SQLEXPRESS" -Cluster $Cluster -ErrorAction Stop) {
-                    Write-PSFMessage -Level Verbose -String 'New-AIPScannerInstall.Message10'
+                    Write-PSFMessage -Level Verbose -String 'New-AIPScannerInstall.Message11'
                 }
                 else {
-                    Write-PSFMessage -Level Verbose -String 'New-AIPScannerInstall.Message11'
+                    Write-PSFMessage -Level Verbose -String 'New-AIPScannerInstall.Message12'
                 }
             }
             catch {
-                Stop-PSFFunction -String 'New-AIPScannerInstall.Message12' -EnableException $EnableException -Cmdlet $PSCmdlet -ErrorRecord $_
+                Stop-PSFFunction -String 'New-AIPScannerInstall.Message13' -EnableException $EnableException -Cmdlet $PSCmdlet -ErrorRecord $_
             }
 
             try {
-                Write-PSFMessage -Level Verbose -String 'New-AIPScannerInstall.Message13'
+                Write-PSFMessage -Level Verbose -String 'New-AIPScannerInstall.Message14'
                 $service = Get-Service | Where-Object Name -eq 'AIPScanner'
                 if($service.Status -eq 'Stopped'){
-                    Write-PSFMessage -Level Verbose -String 'New-AIPScannerInstall.Message14'
+                    Write-PSFMessage -Level Verbose -String 'New-AIPScannerInstall.Message15'
                     Start-Service -Name AIPScanner -ErrorAction Stop
                 }
                 else {
-                    Write-PSFMessage -Level Verbose -String 'New-AIPScannerInstall.Message15'
+                    Write-PSFMessage -Level Verbose -String 'New-AIPScannerInstall.Message16'
                 }
             }
             catch {
-                Stop-PSFFunction -String 'New-AIPScannerInstall.Message16' -EnableException $EnableException -Cmdlet $PSCmdlet -ErrorRecord $_
+                Stop-PSFFunction -String 'New-AIPScannerInstall.Message17' -EnableException $EnableException -Cmdlet $PSCmdlet -ErrorRecord $_
             }
 
         }
         catch {
-            Stop-PSFFunction -String 'New-AIPScannerInstall.Message17' -EnableException $EnableException -Cmdlet $PSCmdlet -ErrorRecord $_
+            Stop-PSFFunction -String 'New-AIPScannerInstall.Message18' -EnableException $EnableException -Cmdlet $PSCmdlet -ErrorRecord $_
         }
     }
 
     end {
-        Write-PSFMessage -Level Host -String 'New-AIPScannerInstall.Message18'
+        Write-PSFMessage -Level Host -String 'New-AIPScannerInstall.Message19'
     }
 }